rename class constants
