+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/utils/ErrorMessage.java
@@ -1,29 +1,25 @@
 package com.kbtg.bootcamp.posttest.utils;
 
-public enum ErrorMessage {
-    VALIDATION_FAILED("Validation Failed"),
-    RESOURCE_NOT_FOUND("Resource not found"),
-    USERID_VALIDATE_LENGTH_MSG("UserId length should be 10 characters"),
-    TICKET_ID_VALIDATE_LENGTH_MSG("TicketId length should be 6 characters"),
-    NO_RESOURCE_FOUND_TO_SELL_BACK("No resource found to sell back"),
-    TICKET_IS_REQUIRED("Ticket is required"),
-    PRICE_IS_REQUIRED("Price is required"),
-    AMOUNT_IS_REQUIRED("Amount is required"),
-    INVALID_TICKET_ID("Invalid tickets id"),
-    INVALID_PRICE_VALUE("Invalid price value"),
-    INVALID_AMOUNT_VALUE("Invalid amount value"),
-    ERROR_OCCURRED_BUY_LOTTERY("Error occurred while buying lottery tickets"),
-    ERROR_TICKET_ALREADY_EXIST("Ticket already exist"),
-    ERROR_TICKET_NOT_FOUND("Lottery Tickets not found"),
-    TICKETS_HAVE_BEEN_SOLD_OUT("Lottery tickets have been sold out");
+public final class ErrorMessage {
 
-    private String message;
-
-    ErrorMessage(String message) {
-        this.message = message;
+    private ErrorMessage() {
     }
 
-    public String getMessage() {
-        return this.message;
-    }
+    public static final String VALIDATION_FAILED = "Validation Failed";
+    public static final String RESOURCE_NOT_FOUND = "Resource not found";
+    public static final String USERID_VALIDATE_LENGTH_MSG = "UserId length should be 10 characters";
+    public static final String TICKET_ID_VALIDATE_LENGTH_MSG = "TicketId length should be 6 characters";
+    public static final String NO_RESOURCE_FOUND_TO_SELL_BACK = "No resource found to sell back";
+    public static final String TICKET_IS_REQUIRED = "Ticket is required";
+    public static final String PRICE_IS_REQUIRED = "Price is required";
+    public static final String AMOUNT_IS_REQUIRED = "Amount is required";
+    public static final String INVALID_TICKET_ID = "Invalid tickets id";
+    public static final String INVALID_PRICE_VALUE = "Invalid price value";
+    public static final String INVALID_AMOUNT_VALUE = "Invalid amount value";
+    public static final String ERROR_OCCURRED_BUY_LOTTERY = "Error occurred while buying lottery tickets";
+    public static final String ERROR_TICKET_ALREADY_EXIST = "Ticket already exist";
+    public static final String ERROR_TICKET_NOT_FOUND = "Lottery Tickets not found";
+    public static final String TICKETS_HAVE_BEEN_SOLD_OUT = "Lottery tickets have been sold out";
+
 }
+
