+++ posttest/src/test/java/com/kbtg/bootcamp/posttest/units/CreateLotteryRequestDtoTest.java
@@ -12,13 +12,13 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Iterator;
 
-import static com.kbtg.bootcamp.posttest.utils.Constants.AMOUNT_IS_REQUIRED;
-import static com.kbtg.bootcamp.posttest.utils.Constants.INVALID_AMOUNT_VALUE;
-import static com.kbtg.bootcamp.posttest.utils.Constants.INVALID_PRICE_VALUE;
-import static com.kbtg.bootcamp.posttest.utils.Constants.INVALID_TICKET_ID;
-import static com.kbtg.bootcamp.posttest.utils.Constants.PRICE_IS_REQUIRED;
-import static com.kbtg.bootcamp.posttest.utils.Constants.TICKET_ID_VALIDATE_LENGTH_MSG;
-import static com.kbtg.bootcamp.posttest.utils.Constants.TICKET_IS_REQUIRED;
+import static com.kbtg.bootcamp.posttest.utils.ErrorMessage.AMOUNT_IS_REQUIRED;
+import static com.kbtg.bootcamp.posttest.utils.ErrorMessage.INVALID_AMOUNT_VALUE;
+import static com.kbtg.bootcamp.posttest.utils.ErrorMessage.INVALID_PRICE_VALUE;
+import static com.kbtg.bootcamp.posttest.utils.ErrorMessage.INVALID_TICKET_ID;
+import static com.kbtg.bootcamp.posttest.utils.ErrorMessage.PRICE_IS_REQUIRED;
+import static com.kbtg.bootcamp.posttest.utils.ErrorMessage.TICKET_ID_VALIDATE_LENGTH_MSG;
+import static com.kbtg.bootcamp.posttest.utils.ErrorMessage.TICKET_IS_REQUIRED;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @ExtendWith(MockitoExtension.class)
